Fix typo in the name of the `DistributedCacheRateLimitCounterHandler` class (#1969) by @williamscs
